Move the verificator Dockerfile and scripts to apps
