+++ apps/rendering/resources/images/requirements.txt
@@ -3,3 +3,8 @@ opencv-python
 OpenEXR==1.3.0
 scikit-image
 scipy
+Pillow==5.1.0
+sklearn
+numpy
+PyWavelets
+pydot
