+++ apps/rendering/resources/scripts/imgmetrics.py
@@ -1,6 +1,26 @@
 import io
 import os
 import json
+import numpy as np
+import sys
+
+from . import metrics
+ 
+
+class MyEncoder(json.JSONEncoder):
+    def default(self, obj):
+        print("There were obj %r" % obj, file=sys.stderr)
+        
+        if isinstance(obj, np.integer):
+            return int(obj)
+        elif isinstance(obj, np.float32):
+            return float(obj)
+        elif isinstance(obj, np.float64):
+            return float(obj)
+        elif isinstance(obj, np.ndarray):
+            return obj.tolist()
+        else:
+            return obj.__dict__
 
 
 class ImgMetrics:
@@ -10,20 +30,36 @@ class ImgMetrics:
     """
 
     def __init__(self, dictionary=None):
-        self.imgCorr = None  # for intellisense
-        self.SSIM_normal = None
-        self.MSE_normal = None
-        self.SSIM_canny = None
-        self.MSE_canny = None
-        self.SSIM_wavelet = None
-        self.MSE_wavelet = None
+        self.ssim = None
+        self.reference_variance = None
+        self.image_variance = None
+        self.ref_edge_factor = None
+        self.comp_edge_factor = None
+        self.edge_difference = None
+        self.wavelet_sym2_base = None
+        self.wavelet_sym2_low = None
+        self.wavelet_sym2_mid = None
+        self.wavelet_sym2_high = None
+        self.wavelet_db4_base = None
+        self.wavelet_db4_low = None
+        self.wavelet_db4_mid = None
+        self.wavelet_db4_high = None
+        self.wavelet_haar_base = None
+        self.wavelet_haar_low = None
+        self.wavelet_haar_mid = None
+        self.wavelet_haar_high = None
+        self.wavelet_haar_freq_x1 = None
+        self.wavelet_haar_freq_x2 = None
+        self.wavelet_haar_freq_x3 = None
+        self.histograms_correlation = None
+        self.max_x_mass_center_distance = None
+        self.max_y_mass_center_distance = None
         self.crop_resolution = None
+        self.variance_difference = None
+
         # ensure that the keys are correct
-        keys = ['imgCorr',
-                'SSIM_normal', 'MSE_normal',
-                'SSIM_canny', 'MSE_canny',
-                'SSIM_wavelet', 'MSE_wavelet',
-                'crop_resolution']
+        keys = ImgMetrics.get_metric_names()
+        keys.append('Label')
 
         for key in keys:
             if key not in dictionary:
@@ -33,9 +69,30 @@ class ImgMetrics:
         for key in dictionary:
             setattr(self, key, dictionary[key])
 
+    @staticmethod
+    def get_metric_classes():
+        available_metrics = [
+            metrics.ssim.MetricSSIM,
+            metrics.psnr.MetricPSNR,
+            metrics.variance.ImageVariance,
+            metrics.edges.MetricEdgeFactor,
+            metrics.wavelet.MetricWavelet,
+            metrics.histograms_correlation.MetricHistogramsCorrelation,
+            metrics.mass_center_distance.MetricMassCenterDistance
+        ]
+        
+        return available_metrics
+
+    @staticmethod
+    def get_metric_names():
+        metric_names = []
+        for metric_class in ImgMetrics.get_metric_classes():
+            metric_names = metric_names + metric_class.get_labels()
+        return metric_names
+
     def to_json(self):
         str_ = json.dumps(self,
-                          default=lambda o: o.__dict__,
+                          cls=MyEncoder,
                           indent=4,
                           sort_keys=True,
                           separators=(',', ': '),
