+++ apps/rendering/resources/scripts/metrics/edges.py
