+++ apps/rendering/resources/scripts/metrics/mass_center_distance.py
