+++ apps/rendering/resources/scripts/metrics/psnr.py
