+++ apps/rendering/resources/scripts/metrics/ssim.py
