+++ apps/rendering/resources/scripts/metrics/variance.py
