--- golem/verificator/docker/blender/images/requirements.txt
@@ -1,10 +1,0 @@
---extra-index-url https://builds.golem.network
-opencv-python
-OpenEXR
-scikit-image
-scipy
-Pillow==5.1.0
-sklearn
-numpy
-PyWavelets
-pydot
